Change the OpenFile class to service class pattern
