+++ main.rb
@@ -1,8 +1,8 @@
-require_relative 'services/open_files'
+require_relative 'services/csv/open_files'
 
 class Main
   def initialize
-    @accounts_csv, @transaction_csv = Services::OpenFiles.new(ARGV).files
+    @accounts_csv, @transaction_csv = Services::CSV::OpenFiles.new(ARGV).files
   end
 end
